+++ activiti-core/activiti-engine/src/test/java/org/activiti/engine/test/bpmn/event/timer/BoundaryTimerNonInterruptingEventTest.java
@@ -338,7 +338,8 @@ public class BoundaryTimerNonInterruptingEventTest extends PluggableActivitiTest
   }
 
   private void moveByMinutes(int minutes) throws Exception {
-    processEngineConfiguration.getClock().setCurrentTime(new Date(processEngineConfiguration.getClock().getCurrentTime().getTime() + ((minutes * 60 * 1000))));
+    processEngineConfiguration.getClock().setCurrentTime(new Date(processEngineConfiguration.getClock().getCurrentTime().getTime() + ((
+      (long) minutes * 60 * 1000))));
   }
 
 }
